+++ src/components/ItaliaTheme/Blocks/BandiSearch/Sidebar.jsx
@@ -12,6 +12,7 @@ import QueryWidget from '@plone/volto/components/manage/Widgets/QueryWidget';
 import upSVG from '@plone/volto/icons/up-key.svg';
 import downSVG from '@plone/volto/icons/down-key.svg';
 import FiltersConfig from 'design-comuni-plone-theme/components/ItaliaTheme/Blocks/BandiSearch/FiltersConfig';
+import config from '@plone/volto/registry';
 
 const messages = defineMessages({
   help: {
@@ -64,7 +65,8 @@ const messages = defineMessages({
   },
   help_filters: {
     id: 'searchBlock_help_filters',
-    defaultMessage: "Questi filtri non verranno visualizzati dall'utente ma consentono di pre filtrare i bandi secondo alcuni criteri.",
+    defaultMessage:
+      "Questi filtri non verranno visualizzati dall'utente ma consentono di pre filtrare i bandi secondo alcuni criteri.",
   },
   text_filter: {
     id: 'searchBlock_text_filter',
@@ -106,9 +108,22 @@ const messages = defineMessages({
     id: 'sort_modified_date',
     defaultMessage: 'Data di modifica',
   },
+  variation: {
+    id: 'search_bandi_variation',
+    defaultMessage: 'Aspetto dei risultati',
+  },
+  display_results_by_default: {
+    id: 'display_results_by_default',
+    defaultMessage: 'Mostra subito i risultati',
+  },
+  display_results_by_default_description: {
+    id: 'display_results_by_default_description',
+    defaultMessage:
+      "Mostra subito i risultati della ricerca, senza che l'utente abbia gia applicato dei filtri",
+  },
 });
 
-const Sidebar = ({ block, data, onChangeBlock, required }) => {
+const Sidebar = ({ block, data, onChangeBlock, required, ...others }) => {
   const intl = useIntl();
 
   /* Accordions active */
@@ -137,6 +152,15 @@ const Sidebar = ({ block, data, onChangeBlock, required }) => {
     ['effective', intl.formatMessage(messages.sort_effective_date)],
     ['modified', intl.formatMessage(messages.sort_modified_date)],
   ];
+
+  const variations = config.blocks.blocksConfig.searchBandi.results_variations;
+  let variations_options = null;
+  if (variations.length > 0) {
+    variations_options = [['default', 'Card']];
+    variations.forEach((v) => {
+      variations_options.push([v.id, v.name]);
+    });
+  }
 
   return (
     <Segment.Group raised>
@@ -188,6 +212,21 @@ const Sidebar = ({ block, data, onChangeBlock, required }) => {
             }}
             choices={filters}
           />
+          <CheckboxWidget
+            id="display_results_by_default"
+            title={intl.formatMessage(messages.display_results_by_default)}
+            description={intl.formatMessage(
+              messages.display_results_by_default_description,
+            )}
+            value={
+              data.display_results_by_default
+                ? data.display_results_by_default
+                : false
+            }
+            onChange={(name, checked) => {
+              onChangeBlock(block, { ...data, [name]: checked });
+            }}
+          />
           <SelectWidget
             id="sort_on"
             title={intl.formatMessage(messages.sort_on)}
@@ -222,6 +261,7 @@ const Sidebar = ({ block, data, onChangeBlock, required }) => {
           />
         </div>
       </Segment>
+
       <Accordion fluid styled className="form">
         <Accordion.Title
           active={activeAccFilters}
@@ -238,12 +278,17 @@ const Sidebar = ({ block, data, onChangeBlock, required }) => {
         <Accordion.Content active={activeAccFilters}>
           <Segment padded>
             <p className="help">{intl.formatMessage(messages.help_filters)}</p>
-            <QueryWidget block={block} onChange={(id, value)=> {
-              onChangeBlock(block, {
-                ...data,
-                [id]: value,
-              });
-            }} id='defaultQuerystring' value={data.defaultQuerystring} />
+            <QueryWidget
+              block={block}
+              onChange={(id, value) => {
+                onChangeBlock(block, {
+                  ...data,
+                  [id]: value,
+                });
+              }}
+              id="defaultQuerystring"
+              value={data.defaultQuerystring}
+            />
           </Segment>
         </Accordion.Content>
       </Accordion>
@@ -288,6 +333,21 @@ const Sidebar = ({ block, data, onChangeBlock, required }) => {
               ['tertiary', intl.formatMessage(messages.tertiary)],
             ]}
           />
+
+          {variations_options?.length > 0 && (
+            <SelectWidget
+              id="variation"
+              title={intl.formatMessage(messages.variation)}
+              value={data.variation}
+              onChange={(id, value) => {
+                onChangeBlock(block, {
+                  ...data,
+                  variation: value,
+                });
+              }}
+              choices={variations_options}
+            />
+          )}
         </Accordion.Content>
       </Accordion>
     </Segment.Group>